Move and rename PostProcessFFT
