+++ vsdenoise/postprocess.py
@@ -1,18 +1,25 @@
 from __future__ import annotations
 
-from typing import Any
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Any, Literal, overload
 
 from vsexprtools import ExprOp, ExprToken, norm_expr
 from vskernels import Bilinear
 from vsmasktools import EdgeDetectT, FDoGTCanny, range_mask
-from vsrgtools import bilateral, box_blur, gauss_blur
+from vsrgtools import RemoveGrainMode, bilateral, box_blur, gauss_blur, removegrain
+from vsrgtools.util import norm_rmode_planes
 from vstools import (
-    CustomIndexError, InvalidColorFamilyError, PlanesT, check_ref_clip, check_variable, flatten_vnodes, get_y,
-    normalize_planes, scale_8bit, scale_value, vs
+    CustomIndexError, CustomIntEnum, FuncExceptT, InvalidColorFamilyError, KwargsT, PlanesT, check_ref_clip,
+    check_variable, fallback, flatten_vnodes, get_y, normalize_planes, scale_8bit, scale_value, vs
 )
 
+from .fft import DFTTest, fft3d
+from .knlm import nl_means
+
 __all__ = [
-    'decrease_size'
+    'decrease_size',
+
+    'PostProcess'
 ]
 
 
@@ -82,3 +89,118 @@ def decrease_size(
     denoise = bilateral(clip, sigmaS, sigmaR, **kwargs)
 
     return clip.std.MaskedMerge(denoise, mask, planes)
+
+
+@dataclass
+class PostProcessConfig:
+    mode: PostProcess
+    kwargs: KwargsT
+
+    _sigma: float | None = None
+    _tr: int | None = None
+    _block_size: int | None = None
+    merge_strength: int = 0
+
+    @property
+    def sigma(self) -> float:
+        sigma = fallback(self._sigma, 1.0)
+
+        if self.mode is PostProcess.DFTTEST:
+            return sigma * 4
+
+        if self.mode is PostProcess.NL_MEANS:
+            return sigma / 2
+
+        return sigma
+
+    @property
+    def tr(self) -> int:
+        if self.mode <= 0:
+            return 0
+
+        tr = fallback(self._tr, 1)
+
+        if self.mode is PostProcess.DFTTEST:
+            return min(tr, 3)
+
+        if self.mode in {PostProcess.FFT3D_MED, PostProcess.FFT3D_HIGH}:
+            return min(tr, 2)
+
+        return tr
+
+    @property
+    def block_size(self) -> int:
+        if self.mode is PostProcess.DFTTEST:
+            from .fft import BackendInfo
+
+            backend_info = BackendInfo.from_param(self.kwargs.pop('plugin', DFTTest.Backend.AUTO))
+
+            if backend_info.resolved_backend.is_dfttest2:
+                return 16
+
+        return fallback(self._block_size, [0, 48, 32, 12, 0][self.mode.value])
+
+    def __call__(self, clip: vs.VideoNode, planes: PlanesT = None, func: FuncExceptT | None = None) -> vs.VideoNode:
+        func = func or self.__class__
+
+        if self.mode is PostProcess.REPAIR:
+            return removegrain(clip, norm_rmode_planes(clip, RemoveGrainMode.MINMAX_AROUND1, planes))
+
+        if self.mode in {PostProcess.FFT3D_MED, PostProcess.FFT3D_HIGH}:
+            return fft3d(clip, func, bw=self.block_size, bh=self.block_size, bt=self.tr * 2 + 1, **self.kwargs)
+
+        if self.mode is PostProcess.DFTTEST:
+            return DFTTest.denoise(
+                clip, self.sigma, tr=self.tr, block_size=self.block_size,
+                planes=planes, **(KwargsT(overlap=int(self.block_size * 9 / 12)) | self.kwargs)  # type: ignore
+            )
+
+        if self.mode is PostProcess.NL_MEANS:
+            return nl_means(
+                clip, self.sigma, self.tr, planes=planes, **(KwargsT(sr=2) | self.kwargs)  # type: ignore
+            )
+
+        return clip
+
+
+class PostProcess(CustomIntEnum):
+    REPAIR = 0
+    FFT3D_HIGH = 1
+    FFT3D_MED = 2
+    DFTTEST = 3
+    NL_MEANS = 4
+
+    if TYPE_CHECKING:
+        from .postprocess import PostProcess
+
+        @overload
+        def __call__(  # type: ignore
+            self: Literal[PostProcess.REPAIR], *, merge_strength: int = 0
+        ) -> PostProcessConfig:
+            ...
+
+        @overload
+        def __call__(  # type: ignore
+            self: Literal[PostProcess.NL_MEANS], *, sigma: float = 1.0, tr: int | None = None,
+            merge_strength: int = 0, **kwargs: Any
+        ) -> PostProcessConfig:
+            ...
+
+        @overload
+        def __call__(
+            self, *, sigma: float = 1.0, tr: int | None = None, block_size: int | None = None,
+            merge_strength: int = 0, **kwargs: Any
+        ) -> PostProcessConfig:
+            ...
+
+        def __call__(
+            self, *, sigma: float = 1.0, tr: int | None = None, block_size: int | None = None,
+            merge_strength: int = 0, **kwargs: Any
+        ) -> PostProcessConfig:
+            ...
+    else:
+        def __call__(
+            self, *, sigma: float = 1.0, tr: int | None = None, block_size: int | None = None,
+            merge_strength: int = 0, **kwargs: Any
+        ) -> PostProcessConfig:
+            return PostProcessConfig(self, kwargs, sigma, tr, block_size, merge_strength)
